LayoutTests/fast/css/webkit-line-clamp-calculated-value-expected.txt

Tests assigning a calculated value to -webkit-line-clamp CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['-webkit-line-clamp'] is ""
testDiv.style['-webkit-line-clamp'] = 'calc(10% * 2)'
PASS testDiv.style['-webkit-line-clamp'] is "calc(20%)"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "20%"
testDiv.style['-webkit-line-clamp'] = 'calc(2 * 3)'
PASS testDiv.style['-webkit-line-clamp'] is "calc(6)"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "6"
testDiv.style['-webkit-line-clamp'] = 'calc(-2%)'
PASS testDiv.style['-webkit-line-clamp'] is "calc(-2%)"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "0%"
testDiv.style['-webkit-line-clamp'] = 'calc(-2)'
PASS testDiv.style['-webkit-line-clamp'] is "calc(-2)"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "1"
testDiv.style['-webkit-line-clamp'] = 'calc(10 + (sign(2cqw - 10px) * 5))'
PASS testDiv.style['-webkit-line-clamp'] is "calc(10 + (5 * sign(2cqw - 10px)))"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "5"
testDiv.style['-webkit-line-clamp'] = 'calc(10% + (sign(2cqw - 10px) * 5%))'
PASS testDiv.style['-webkit-line-clamp'] is "calc(10% + (5% * sign(2cqw - 10px)))"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "5%"
PASS successfullyParsed is true

TEST COMPLETE
